+++ packages/components/inline-message/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sl-design-system/inline-message
 
+## 0.2.0
+
+### Minor Changes
+
+- [#983](https://github.com/sl-design-system/components/pull/983) [`1d478e1`](https://github.com/sl-design-system/components/commit/1d478e161dfdc0b249fd0ca8dcb10a6efbd1ec84) - The title is no longer the default slot; this is now a named slot(`title`). The default slot is now the main content.
+
+### Patch Changes
+
+- Updated dependencies [[`216d62e`](https://github.com/sl-design-system/components/commit/216d62eb5a16277b4ea2767ea0530e570bf40abf), [`216d62e`](https://github.com/sl-design-system/components/commit/216d62eb5a16277b4ea2767ea0530e570bf40abf), [`216d62e`](https://github.com/sl-design-system/components/commit/216d62eb5a16277b4ea2767ea0530e570bf40abf)]:
+  - @sl-design-system/shared@0.2.4
+  - @sl-design-system/button@0.0.20
+
 ## 0.1.3
 
 ### Patch Changes