+++ packages/themes/neon/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sanomalearning/slds-neon
 
+## 0.1.4
+
+### Patch Changes
+
+- [#985](https://github.com/sl-design-system/components/pull/985) [`9c88751`](https://github.com/sl-design-system/components/commit/9c887516bacb0a335d8d2283e47d9a78f359a334) - Added tokens for Tab component
+
 ## 0.1.3
 
 ### Patch Changes